fix bug for parser

+++ src/builder/python.sh
@@ -3,7 +3,9 @@
 
 python_build() {
   pip install setuptools wheel -i https://mirrors.aliyun.com/pypi/simple/
-  pip install -r requirements.txt -i https://mirrors.aliyun.com/pypi/simple/
+  if [ ! -f "requirements.txt " ]; then
+    pip install -r requirements.txt -i https://mirrors.aliyun.com/pypi/simple/
+  fi
   python3 setup.py bdist_wheel
 }
